+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 ## Unreleased
+
+- Support decoding types as below:
+    - primitive types like `str`, `int`, `float`, `bool` and `None`.
+    -`Union` and `Optional`.
+    - homogeneous and heterogeneous `Tuple`
+    - variable-length `Tuple`.
+    - non-generic dataclasses.